requirements.txt
flask-bower
flask-login
Flask-Navigation
Flask-Babel
WTForms
flask_migrate